requirements.txt
gunicorn==19.7.1
dj-database-url==0.4.2
dj-email-url==0.0.10
django==1.11
django-debug-toolbar==1.8
django-lazysignup==1.1.2
django-material==1.1.0
requests==2.18.4
whitenoise==3.3.1
yapf==0.19.0
